Basic Algorithm working for 2048
Current functionality includes; Cube movement when a directional button is clicked, cube merging(Adding) when a cube is next moved into a cube of the same value, spawning of a new cube at the end of a move. Things to add; win/lose conditions, scoring system, post score data to webserver (store score in database), add new leaderboard page to website for 2048.